requirements.txt
# 데이터 처리 및 분석
pandas
numpy
scipy

# LLM 및 RAG (Retrieval-Augmented Generation) 관련
requests
qdrant-client
langchain-text-splitters
openai

# HTML/XML 파싱 및 웹 스크래핑
beautifulsoup4

# 클러스터링
hdbscan

# 기타 유틸리티
tqdm # 진행률 표시줄
python-magic # 파일 형식 감지 (import magic)
python-docx # DOCX 파일 처리 (import docx)
PyPDF2 # PDF 처리 (import PyPDF2)
pdfplumber # PDF 처리 (import pdfplumber)
pytesseract # OCR (import pytesseract)
nltk # 자연어 처리
regex # 정규 표현식
mysql-connector-python # MySQL DB 연결

# UI 관련
PySide6
#python -m venv venv
#source venv/bin/activate
#window : .\venv\Scripts\activate